backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
dependency-injector>=4.41
pydantic-settings>=2.0.0
openai>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
jsonschema>=4.0.0

# Enhanced MCP functionality
pydantic>=2.0.0
psutil>=5.9.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
respx>=0.20.0
